go/oasis-node/cmd/storage: Factor pruning command to separate file
